reformat javaconcept take 2
